requirements.txt
PyQt5>=5.15.0
pyserial>=3.5
pymodbus>=3.0.0
numpy>=1.21.0
matplotlib>=3.5.0
reportlab>=3.6.0
fastapi>=0.85.0
uvicorn>=0.18.0
websockets>=10.0
